pyproject.toml
[tool.poetry]
name = "image-remover"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dearpygui = "^1.11.1"
pillow = "^10.4.0"
pydantic = "^2.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
